+++ NULL Engine.sln
@@ -1,9 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 2013 for Windows Desktop
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NULL Engine", "NULL Engine/NULL Engine.vcxproj", "{746CC4C3-787F-4B0E-AA66-E388FE3FF4F6}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EngineCore", "NULL Engine\EngineCore.vcxproj", "{746CC4C3-787F-4B0E-AA66-E388FE3FF4F6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ConfigurationFiles", "ConfigurationFiles", "{B6DAA82C-6554-4C25-B815-1B01401E59B2}"
+	ProjectSection(SolutionItems) = preProject
+		NULL Engine\Resources\Engine\Configuration\configuration.JSON = NULL Engine\Resources\Engine\Configuration\configuration.JSON
+		NULL Engine\Resources\Engine\Configuration\default_configuration.JSON = NULL Engine\Resources\Engine\Configuration\default_configuration.JSON
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,5 +24,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {75F7B616-59D8-4F0F-94AD-00294127221D}
 	EndGlobalSection
 EndGlobal
