--- NULL Engine/Intermediate/NULL Engine.exe.recipe
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>C:\Users\angot\Documents\GitHub\Projecte3\NULL Engine\Build\NULL Engine.exe</ProjectOutputs>
-  <ContentFiles></ContentFiles>
-  <SatelliteDlls></SatelliteDlls>
-  <NonRecipeFileRefs></NonRecipeFileRefs>
-</Project>
